clean up codebase

+++ db/repository/user.sql.go
@@ -60,7 +60,7 @@ DELETE FROM users
 WHERE id = $1
 `
 
-func (q *Queries) DeleteUser(ctx context.Context, id int32) error {
+func (q *Queries) DeleteUser(ctx context.Context, id int64) error {
 	_, err := q.db.Exec(ctx, deleteUser, id)
 	return err
 }
@@ -87,7 +87,7 @@ SELECT id, oauth_sub, email, name FROM users
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetUserByID(ctx context.Context, id int32) (User, error) {
+func (q *Queries) GetUserByID(ctx context.Context, id int64) (User, error) {
 	row := q.db.QueryRow(ctx, getUserByID, id)
 	var i User
 	err := row.Scan(
@@ -123,7 +123,7 @@ WHERE id = $1
 `
 
 type UpdateUserParams struct {
-	ID   int32
+	ID   int64
 	Name string
 }
